refactor: remove contract slice and update store imports
