+++ package.json
@@ -12,6 +12,7 @@
         "@types/node": "^22.10.10",
         "@types/react": "^19.0.8",
         "@types/react-dom": "^19.0.3",
+        "@types/react-redux": "^7.1.34",
         "axios": "^1.7.9",
         "eslint": "^9.20.1",
         "eslint-config-prettier": "^10.0.1",
