+++ resources/ts/main.tsx
@@ -4,7 +4,7 @@ import App from './App';
 import { PrimeReactProvider } from 'primereact/api';
 import '@resources/css/app.css';
 import { Provider } from 'react-redux';
-import store from './store/store';
+import store from './config/stores/Userstore';
 
 import '@/config/i18n';
 import { AuthProvider } from '@/contexts/AuthContext';
@@ -12,7 +12,7 @@ import { ToastProvider } from '@/contexts/ToastContext';
 
 const rootElement = document.getElementById('app');
 if (rootElement) {
-  const root = ReactDOM.createRoot(rootElement); // This is correct for React 18+
+  const root = ReactDOM.createRoot(rootElement);
   root.render(
     <React.StrictMode>
       <Provider store={store}>
