+++ resources/ts/pages/Admin/WorkOrders.tsx
@@ -7,7 +7,7 @@ import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
 import CrudPanel from '@/components/Admin/CrudPanel';
 import { useSelector } from 'react-redux';
-import { RootState } from '@/store/store';
+import { RootState } from '@/config/stores/Userstore';
 
 interface WorkOrder {
   id: number;
